[C++] add explicit cast to prevent warnings.

+++ aeron-client/src/main/cpp/util/Index.h
@@ -29,7 +29,7 @@ typedef std::int32_t index_t;
 
 inline static index_t convertSizeToIndex(size_t size)
 {
-    if (size > std::numeric_limits<index_t>::max())
+    if (size > static_cast<size_t>(std::numeric_limits<index_t>::max()))
     {
         return std::numeric_limits<index_t>::max();
     }
